move broadcasted lt tests into sum lazy tensor
